refactor: Go out JSON RPC mock datas.
